+++ Core/WebActionContext.cs
@@ -209,7 +209,7 @@ namespace Greatbone.Core
             return entity as M;
         }
 
-        public async Task<D> ReadObjectAsync<D>(ushort proj = 0) where D : IData, new()
+        public async Task<D> ReadObjectAsync<D>(int proj = 0) where D : IData, new()
         {
             if (entity == null && count == -1) // if not yet parse and read
             {
@@ -236,7 +236,7 @@ namespace Greatbone.Core
             return src.ToObject<D>(proj);
         }
 
-        public async Task<D[]> ReadArrayAsync<D>(ushort proj = 0) where D : IData, new()
+        public async Task<D[]> ReadArrayAsync<D>(int proj = 0) where D : IData, new()
         {
             if (entity == null && count == -1) // if not yet parse and read
             {
@@ -258,7 +258,7 @@ namespace Greatbone.Core
             return (entity as IDataInput)?.ToArray<D>(proj);
         }
 
-        public async Task<List<D>> ReadListAsync<D>(ushort proj = 0) where D : IData, new()
+        public async Task<List<D>> ReadListAsync<D>(int proj = 0) where D : IData, new()
         {
             if (entity == null && count == -1) // if not yet parse and read
             {
@@ -363,7 +363,7 @@ namespace Greatbone.Core
 
         static readonly TypeInfo ListType = typeof(List<IData>).GetTypeInfo();
 
-        public void ReplyJson(int status, object data, ushort proj = 0, bool? pub = null, int maxage = 60)
+        public void ReplyJson(int status, object data, int proj = 0, bool? pub = null, int maxage = 60)
         {
             TypeInfo typ = data.GetType().GetTypeInfo();
 
@@ -389,7 +389,7 @@ namespace Greatbone.Core
             MaxAge = maxage;
         }
 
-        public void ReplyXml(int status, object dat, ushort proj = 0, bool? pub = null, int maxage = 60)
+        public void ReplyXml(int status, object dat, int proj = 0, bool? pub = null, int maxage = 60)
         {
         }
